Backend/data-processing-service/processing-service/requirements.txt
confluent-kafka>=2.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic==2.6.1
marshmallow>=3.20.0
prometheus-client>=0.19.0
aiohttp>=3.9.0
motor>=3.3.0
numpy
langchain-community